@@ -6,7 +6,7 @@ import warnings
 from collections import defaultdict
 from datetime import datetime
 from enum import Enum
-from typing import Any, Callable, DefaultDict, List, Mapping, Optional, Tuple, Type, Union
+from typing import Any, Callable, DefaultDict, Mapping, Type
 
 from torch.optim import Optimizer
 
@@ -325,16 +325,16 @@ class OutputHandler(BaseOutputHandler):
     def __init__(
         self,
         tag: str,
-        metric_names: Optional[Union[List[str], str]] = None,
-        output_transform: Optional[Callable] = None,
-        global_step_transform: Optional[Callable[[Engine, Union[str, Events]], int]] = None,
-        state_attributes: Optional[List[str]] = None,
+        metric_names: list[str] | str | None = None,
+        output_transform: Callable | None = None,
+        global_step_transform: Callable[[Engine, str | Events], int] | None = None,
+        state_attributes: list[str] | None = None,
     ):
         super(OutputHandler, self).__init__(
             tag, metric_names, output_transform, global_step_transform, state_attributes
         )
 
-    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: Union[str, Events]) -> None:
+    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: str | Events) -> None:
         if not isinstance(logger, ClearMLLogger):
             raise RuntimeError("Handler OutputHandler works only with ClearMLLogger")
 
@@ -392,10 +392,10 @@ class OptimizerParamsHandler(BaseOptimizerParamsHandler):
             )
     """
 
-    def __init__(self, optimizer: Optimizer, param_name: str = "lr", tag: Optional[str] = None):
+    def __init__(self, optimizer: Optimizer, param_name: str = "lr", tag: str | None = None):
         super(OptimizerParamsHandler, self).__init__(optimizer, param_name, tag)
 
-    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: Union[str, Events]) -> None:
+    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: str | Events) -> None:
         if not isinstance(logger, ClearMLLogger):
             raise RuntimeError("Handler OptimizerParamsHandler works only with ClearMLLogger")
 
@@ -489,7 +489,7 @@ class WeightsScalarHandler(BaseWeightsScalarHandler):
         optional argument `whitelist` added.
     """
 
-    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: Union[str, Events]) -> None:
+    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: str | Events) -> None:
         if not isinstance(logger, ClearMLLogger):
             raise RuntimeError("Handler WeightsScalarHandler works only with ClearMLLogger")
 
@@ -579,7 +579,7 @@ class WeightsHistHandler(BaseWeightsHandler):
         optional argument `whitelist` added.
     """
 
-    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: Union[str, Events]) -> None:
+    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: str | Events) -> None:
         if not isinstance(logger, ClearMLLogger):
             raise RuntimeError("Handler 'WeightsHistHandler' works only with ClearMLLogger")
 
@@ -675,7 +675,7 @@ class GradsScalarHandler(BaseWeightsScalarHandler):
         optional argument `whitelist` added.
     """
 
-    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: Union[str, Events]) -> None:
+    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: str | Events) -> None:
         if not isinstance(logger, ClearMLLogger):
             raise RuntimeError("Handler GradsScalarHandler works only with ClearMLLogger")
 
@@ -765,7 +765,7 @@ class GradsHistHandler(BaseWeightsHandler):
             optional argument `whitelist` added.
     """
 
-    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: Union[str, Events]) -> None:
+    def __call__(self, engine: Engine, logger: ClearMLLogger, event_name: str | Events) -> None:
         if not isinstance(logger, ClearMLLogger):
             raise RuntimeError("Handler 'GradsHistHandler' works only with ClearMLLogger")
 
@@ -828,9 +828,9 @@ class ClearMLSaver(DiskSaver):
 
     def __init__(
         self,
-        logger: Optional[ClearMLLogger] = None,
-        output_uri: Optional[str] = None,
-        dirname: Optional[str] = None,
+        logger: ClearMLLogger | None = None,
+        output_uri: str | None = None,
+        dirname: str | None = None,
         *args: Any,
         **kwargs: Any,
     ):
@@ -850,7 +850,7 @@ class ClearMLSaver(DiskSaver):
         if "atomic" not in kwargs:
             kwargs["atomic"] = False
 
-        self._checkpoint_slots: DefaultDict[Union[str, Tuple[str, str]], List[Any]] = defaultdict(list)
+        self._checkpoint_slots: DefaultDict[str | tuple[str, str], list[Any]] = defaultdict(list)
 
         super(ClearMLSaver, self).__init__(dirname=dirname, *args, **kwargs)  # type: ignore[misc]
 
@@ -885,11 +885,11 @@ class ClearMLSaver(DiskSaver):
         def __init__(
             self,
             callback_type: Type[Enum],
-            slots: List,
+            slots: list,
             checkpoint_key: str,
             filename: str,
             basename: str,
-            metadata: Optional[Mapping] = None,
+            metadata: Mapping | None = None,
         ) -> None:
             self._callback_type = callback_type
             self._slots = slots
@@ -930,7 +930,7 @@ class ClearMLSaver(DiskSaver):
 
             return model_info
 
-    def __call__(self, checkpoint: Mapping, filename: str, metadata: Optional[Mapping] = None) -> None:
+    def __call__(self, checkpoint: Mapping, filename: str, metadata: Mapping | None = None) -> None:
         try:
             from clearml.binding.frameworks import WeightsFileHandler
         except ImportError:
@@ -970,7 +970,7 @@ class ClearMLSaver(DiskSaver):
             WeightsFileHandler.remove_post_callback(post_cb_id)
 
     @idist.one_rank_only()
-    def get_local_copy(self, filename: str) -> Optional[str]:
+    def get_local_copy(self, filename: str) -> str | None:
         """Get artifact local copy.
 
         .. warning::
